Move telemetry action to DefaultStoresManager
